added products endpoint, tests and products by section_id

+++ backend/src/psql/init.sql
@@ -14,7 +14,7 @@ CREATE TABLE "Product" (
     name VARCHAR(100),
     image VARCHAR(500),
     section_id INTEGER REFERENCES "Section"(id),
-    price INTEGER,
+    price DOUBLE PRECISION,
     ingredients VARCHAR(500)
 );
